[AAP-80609] Refactor migrate_service_data into modular mixins
Split the 1783-line monolith into a thin Command class (402 lines)
that inherits from focused mixins in _migrate_service_data/:

- cursor_store.py: CursorStore (PK-based crash recovery)
- logging_mixin.py: _configure_logging, _log, _log_progress
- legacy_authenticators.py: delete_legacy_authenticators
- resource_migration.py: migrate_resource, reconcile, sync payloads
- role_assignments.py: pagination, assignment creation, drift check
- superuser_sync.py: superuser consistency across services
- user_merge.py: partially migrated user merging

Pure refactor — no behavior changes. All existing tests should pass
unchanged since Command inherits all mixin methods via MRO.
